Show progress indicator when creating an order
